vimrc: replace vim-sleuth with detectindent
Seems vim-sleuth is a bit slow when opening a file which
doesn't have a clear indentation style and there isn't
any neighbour to check, so it wastes a lot of time going
back in the hierarchy.
Probably I can live just checking the file itself and not
its neighbours, since it's not like I'm creating new files
every minute.