vim: set undercurl start and end control sequences
This allows supported terminal to use the undercurl capability which can
be used in some highlight contexts (e.g. gruvbox use it to highlight the
SpellBad class).

Seems this configuration does not harm terminals which don't support
undercurl (e.g. both screen and tmux fallbacks to underline when this
option is set).