lazyvim: do not autoselect completion suggestions
It is annoying when you have to insert a new-line and instead nvim-cmp
decides that it has to insert a completion suggestion.
Set the <CR> binding so that it completes a suggestion only if this has
been explicitly selected.

Also add a new <C-j> binding with the old <CR> behavior, just in case.