lazyvim: add optional to customized plugin
When a plugin is marked as optional in user config, it is not loaded if
it's not defined somewhere else as well.
Setting optional flag for pre-defined plugins which have been
customized, allows to not load them when the author of the LazyVim
distribution decides to remove them.

This is a preparation for the upgrade of LazyVim, which is going to drop
nvim-cmp and telescope.