lazyvim/nvim-lspconfig: add custom root detection for tinymist lsp
Default nvim-lspconfig for tinymist is to set root_dir variable based on
looking for the .git folder.
Often the typst document is not the only thing present in a git
repository, so it would probably be located in a subdir.

Change the tinymist root_dir function to use LazyVim.root.get(), which
is already configured to watch several root markers (e.g. .project, .git,
.svn, etc).